requirements_mac.txt
aicsimageio==4.14.0
aiobotocore==2.5.4
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
asciitree==0.3.3
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.30.1
azure-storage-blob==12.19.1
bitstruct==8.19.0
black==24.2.0
botocore==1.31.17
bsdiff4==1.2.4
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
Cython
dask==2023.12.1
detools==0.53.0
diff-match-patch==20230430
dill==0.3.8
distributed==2023.12.1
elementpath==4.1.5
exceptiongroup==1.2.0
fasteners==0.19
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.6.0
heatshrink2==0.13.0
humanfriendly==10.0
idna==3.6
imagecodecs==2023.9.18
imageio==2.33.1
importlib-metadata==7.0.0
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
lazy-import==0.2.2
lazy_loader==0.3
locket==1.0.0
lxml==4.9.3
lz4==4.3.3
macholib==1.16.3
MarkupSafe==2.1.3
matplotlib==3.8.2
memory-profiler==0.61.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
natsort==8.4.0
networkx==3.2.1
numcodecs==0.12.1
numpy==1.26.2
ome-types==0.4.5
ome-zarr==0.8.3
onnxruntime==1.16.3
onnxruntime-silicon==1.15.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
openslide-python==1.3.1
packaging==23.2
pandas==2.1.4
parameterized==0.9.0
partd==1.4.1
pathos==0.3.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.2.0
pluggy==1.4.0
pox==0.3.4
ppft==1.7.6.8
protobuf==4.25.1
psutil==5.9.7
PyAutoGUI==0.9.54
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.2
pydantic-compat==0.1.2
pydantic_core==2.14.5
pyelftools==0.31
PyGetWindow==0.0.9
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyobjc-framework-ScriptingBridge==10.1
pyometiff==1.0.0
pyparsing==3.1.1
pyperclip==1.8.2
PyQt6==6.7.0
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
PyRect==0.2.0
PyScreeze==0.1.30
pytest==8.1.1
pytest-qt==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2023.3.post1
pyvips==2.2.2
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
resource-backed-dask-array==0.1.0
roifile==2024.5.24
rubicon-objc==0.4.7
s3fs==2023.6.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
screeninfo==0.8.1
seaborn==0.13.0
securesystemslib==0.31.0
shapely==2.0.2
six==1.16.0
sortedcontainers==2.4.0
supervision==0.17.1
sympy==1.12
tblib==3.0.0
threadpoolctl==3.2.0
tifffile==2023.2.28
tiffslide==2.4.0
tomli==2.0.1
toolz==0.12.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
tuf==4.0.0
tufup==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
webcolors==1.13
wrapt==1.16.0
xarray==2023.12.0
xmlschema==2.5.0
xmltodict==0.13.0
xsdata==23.8
yarl==1.9.4
zarr==2.15.0
zict==3.0.0
zipp==3.17.0
zstandard==0.22.0
azure-storage-blob